backend/requirements.txt
fastapi
uvicorn
typecast-python
python-multipart
python-dotenv
